+++ entities/newsletter/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @brains/newsletter-entity
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/plugins@0.2.0-alpha.123
+  - @brains/contracts@0.2.0-alpha.123
+  - @brains/ui-library@0.2.0-alpha.123
+  - @brains/utils@0.2.0-alpha.123
+  - @brains/entity-service@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes