+++ entities/social-media/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @brains/social-media
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/plugins@0.2.0-alpha.123
+  - @brains/atproto-contracts@0.2.0-alpha.123
+  - @brains/contracts@0.2.0-alpha.123
+  - @brains/utils@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes