+++ shared/document/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brains/document
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.123
+  - @brains/entity-service@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes