+++ shared/media-renderer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @brains/media-renderer
 
+## 0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ## 0.2.0-alpha.121