+++ shared/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brains/test-utils
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes