+++ shell/conversation-service/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/conversation-service
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.123
+  - @brains/utils@0.2.0-alpha.123
+  - @brains/messaging-service@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes