+++ shell/templates/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brains/templates
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/content-formatters@0.2.0-alpha.123
+  - @brains/utils@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes