+++ sites/rizom/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @brains/site-rizom
 
+## 0.2.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/plugins@0.2.0-alpha.123
+  - @brains/atproto-contracts@0.2.0-alpha.123
+  - @brains/site-composition@0.2.0-alpha.123
+  - @brains/ui-library@0.2.0-alpha.123
+  - @brains/utils@0.2.0-alpha.123
+  - @brains/templates@0.2.0-alpha.123
+
 ## 0.2.0-alpha.122
 
 ### Patch Changes